+++ examples/boids/src/main.rs
@@ -283,6 +283,7 @@ impl wgpu_example::framework::Example for Example {
             label: None,
             color_attachments: &color_attachments,
             depth_stencil_attachment: None,
+            occlusion_query_set: None,
         };
 
         // get command encoder