+++ examples/conservative-raster/src/main.rs
@@ -273,6 +273,7 @@ impl wgpu_example::framework::Example for Example {
                     },
                 })],
                 depth_stencil_attachment: None,
+                occlusion_query_set: None,
             });
 
             rpass.set_pipeline(&self.pipeline_triangle_conservative);
@@ -292,6 +293,7 @@ impl wgpu_example::framework::Example for Example {
                     },
                 })],
                 depth_stencil_attachment: None,
+                occlusion_query_set: None,
             });
 
             rpass.set_pipeline(&self.pipeline_upscale);