+++ examples/skybox/src/main.rs
@@ -439,6 +439,7 @@ impl wgpu_example::framework::Example for Skybox {
                     }),
                     stencil_ops: None,
                 }),
+                occlusion_query_set: None,
             });
 
             rpass.set_bind_group(0, &self.bind_group, &[]);