+++ examples/stencil-triangles/src/main.rs
@@ -211,6 +211,7 @@ impl wgpu_example::framework::Example for Triangles {
                         store: true,
                     }),
                 }),
+                occlusion_query_set: None,
             });
 
             rpass.set_stencil_reference(1);