+++ examples/texture-arrays/src/main.rs
@@ -383,6 +383,7 @@ impl wgpu_example::framework::Example for Example {
                 },
             })],
             depth_stencil_attachment: None,
+            occlusion_query_set: None,
         });
 
         rpass.set_pipeline(&self.pipeline);