+++ tests/tests/shader_primitive_index/mod.rs
@@ -176,6 +176,7 @@ fn pulling_common(
         .create_command_encoder(&wgpu::CommandEncoderDescriptor::default());
     {
         let mut rpass = encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
+            label: None,
             color_attachments: &[Some(wgpu::RenderPassColorAttachment {
                 ops: wgpu::Operations {
                     load: wgpu::LoadOp::Clear(wgpu::Color::WHITE),
@@ -185,7 +186,7 @@ fn pulling_common(
                 view: &color_view,
             })],
             depth_stencil_attachment: None,
-            label: None,
+            occlusion_query_set: None,
         });
 
         rpass.set_pipeline(&pipeline);