+++ wgpu-core/src/device/trace.rs
@@ -181,6 +181,7 @@ pub enum Command {
         base: crate::command::BasePass<crate::command::RenderCommand>,
         target_colors: Vec<Option<crate::command::RenderPassColorAttachment>>,
         target_depth_stencil: Option<crate::command::RenderPassDepthStencilAttachment>,
+        occlusion_query_set_id: Option<id::QuerySetId>,
     },
 }